CDTT-472 Merge problem, move CaseEvent to domian.model
