+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/RespondentEventPublisher.java
@@ -1,6 +1,6 @@
 package uk.gov.ons.ctp.integration.rhsvc.message;
 
-import uk.gov.ons.ctp.integration.rhsvc.domain.CaseEvent;
+import uk.gov.ons.ctp.integration.rhsvc.domain.model.CaseEvent;
 
 /**
  * Service responsible for the publication of respondent events to the Response Management System.
