+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/CaseEventReceiverImpl.java
@@ -3,7 +3,7 @@ package uk.gov.ons.ctp.integration.rhsvc.message.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.integration.annotation.MessageEndpoint;
 import org.springframework.integration.annotation.ServiceActivator;
-import uk.gov.ons.ctp.integration.rhsvc.domain.CaseEvent;
+import uk.gov.ons.ctp.integration.rhsvc.domain.model.CaseEvent;
 import uk.gov.ons.ctp.integration.rhsvc.message.CaseEventReceiver;
 import uk.gov.ons.ctp.integration.rhsvc.message.RespondentEventPublisher;
